+++ tests/testthat/test-roxytest-testexamples-best_practice.R
@@ -2,7 +2,7 @@
 
 context("File R/best_practice.R: @testexamples")
 
-test_that("[unknown alias] @ L192", {
+test_that("[unknown alias] @ L197", {
   
   data("simulated_catalogs")
   
@@ -36,6 +36,10 @@ test_that("[unknown alias] @ L192", {
   # Just pick up the solution you want
   obj_s8 <- bp_get_sig_obj(e1, 8)
   
+  # Track the reconstructed profile similarity
+  rec_sim <- get_sig_rec_similarity(obj_s8, t(simulated_catalogs$set1))
+  rec_sim
+  
   # After extraction, you can assign the signatures
   # to reference COSMIC signatures
   # More see ?get_sig_similarity
@@ -47,9 +51,8 @@ test_that("[unknown alias] @ L192", {
   
   # You already got the activities of signatures
   # in obj_suggested, however, you can still
-  # try to optimize the result by removing a signature
-  # in a sample if it contributes <0.01 similarity
-  # For more controls, see its documentation
+  # try to optimize the result.
+  # NOTE: the optimization step may not truly optimize the result!
   expo <- bp_attribute_activity(e1, return_class = "data.table")
   expo$abs_activity
   
@@ -84,6 +87,7 @@ test_that("[unknown alias] @ L192", {
   expect_is(obj_suggested, "Signature")
   expect_is(obj_s8, "Signature")
   expect_is(sim, "list")
+  expect_is(rec_sim, "data.table")
   expect_is(expo, "list")
   expect_is(rank_score, "data.frame")
   expect_is(stats, "list")